Don't break out of superscript on negation, regardless of subtraction behavior.
Summary:
Previously, if you included '-' in `MathQuill.addCharsThatBreakOutOfSupSub()`
it would prevent you from being able to enter negative exponents. This is a
shame because the behavior is otherwise desirable - at all other times, you
want to break out of a super/subscript by default when encountering subtraction
that isn't wrapped in parentheses.

This fixes the behavior by always allowing a (single) negative sign to be
inserted as the first character inside a super/subscript.

Addresses some of T2730

Test Plan:
In Perseus' MathInput, verified the following:
1) Typing 'x^a-b' results in (x^a) - b
2) Typing 'x^-a-b' results in (x^-a) - b
3) Typing 'x^a', moving left once, then typing '-' results in (x^-a)

Reviewers: alpert

Reviewed By: alpert

Subscribers: michelle, jack

Differential Revision: http://phabricator.khanacademy.org/D11366