Fixed tab-trap appearing in certain circumstances - Fixes T3686
Summary:
Mathquill keeps track of a boolean `shouldBeSelected` based on the
input to a `select` method, which determines if the textarea _should_
be selected and, if it isn't, reselects it.

There is a bug where one can select the text in the box, continue
typing, and not be able to escape with the tab key. This is because
the combination of selecting all text and immediately typing does not
reset the `shouldBeSelected` variable as it should. Instead, this
variable only updates if the user selects the contents and hits
"backspace."

The fix for this was simple, we need to invoke `selectionChanged()`
not only when the selection is deleted, but when it is replaced.

Test Plan:
I did the following in Perseus (thanks @aria for providing the repro
steps)

. Create two expression inputs
. In the first, type something
. Select all (Ctrl/Cmd-A)
. Type again
. Hit the tab key
. See you are directed to the next expression input as expected

Reviewers: emily, aria, john

Reviewed By: john

Subscribers: john, alex, aria, #a11y

Maniphest Tasks: T3686

Differential Revision: https://phabricator.khanacademy.org/D16962